Split chat approval persistence tests into submodule
